Update TypeScript configuration and modify landing page redirect
- Added "composite" option to tsconfig for desktop, native, and web apps.
- Updated references in desktop, native, and web tsconfig files to include shared and UI packages.
- Excluded additional directories in root tsconfig.
- Changed landing page redirect from '/dashboard' to '/spaces'.
- Removed unused dashboard page component.
