+++ apps/desktop/tsconfig.json
@@ -9,6 +9,7 @@
     "allowImportingTsExtensions": true,
     "verbatimModuleSyntax": true,
     "noEmit": true,
+    "composite": true,
 
     "skipLibCheck": true,
     "strict": true,
@@ -24,6 +25,10 @@
     }
   },
   "include": ["src/**/*.ts", "src/**/*.tsx"],
-  "exclude": ["node_modules", "dist", "src-tauri/target"],
-  "references": [{ "path": "../server" }]
+  "exclude": ["node_modules", "dist", "src-tauri/target", "src-tauri/gen"],
+  "references": [
+    { "path": "../server" },
+    { "path": "../../packages/shared" },
+    { "path": "../../packages/ui" }
+  ]
 }
