+++ apps/web/tsconfig.json
@@ -14,6 +14,7 @@
     "verbatimModuleSyntax": true,
     "jsx": "preserve",
     "incremental": true,
+    "composite": true,
     "plugins": [
       {
         "name": "next"
@@ -30,5 +31,9 @@
     ".next/types/**/*.ts"
   ],
   "exclude": ["node_modules", "dist", "out", ".next"],
-  "references": [{ "path": "../server" }]
+  "references": [
+    { "path": "../server" },
+    { "path": "../../packages/shared" },
+    { "path": "../../packages/ui" }
+  ]
 }
